// zwangine-designer/src/core/model/workflow-definition.ts
export class ErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'errorHandler';
    id?: string = 'errorHandler-' + uuidv4().substring(0,4);
    deadLetterChannel?: DeadLetterChannelDefinition;
    defaultErrorHandler?: DefaultErrorHandlerDefinition;
    jtaTransactionErrorHandler?: JtaTransactionErrorHandlerDefinition;
    noErrorHandler?: NoErrorHandlerDefinition;
    refErrorHandler?: RefErrorHandlerDefinition | string;
    springTransactionErrorHandler?: SpringTransactionErrorHandlerDefinition;
    public constructor(init?: Partial<ErrorHandlerDefinition>) {
        super('ErrorHandlerDefinition');
        Object.assign(this, init);
    }
}
// ...
export class FromDefinition extends WorkflowElement {
    stepName?: string = 'from';
    id?: string = 'from-' + uuidv4().substring(0,4);
    description?: string;
    uri: string = '';
    variableReceive?: string;
    parameters?: any = {};
    steps: WorkflowElement[] = [];
    public constructor(init?: Partial<FromDefinition>) {
        super('FromDefinition');
        Object.assign(this, init);
    }
}
// ...
export class InterceptDefinition extends WorkflowElement {
    stepName?: string = 'intercept';
    id?: string = 'intercept-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptDefinition>) {
        super('InterceptDefinition');
        Object.assign(this, init);
    }
}

export class InterceptFromDefinition extends WorkflowElement {
    stepName?: string = 'interceptFrom';
    id?: string = 'interceptFrom-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    uri?: string;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptFromDefinition>) {
        super('InterceptFromDefinition');
        Object.assign(this, init);
    }
}

export class InterceptSendToEndpointDefinition extends WorkflowElement {
    stepName?: string = 'interceptSendToEndpoint';
    id?: string = 'interceptSendToEndpoint-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    uri: string = '';
    skipSendToOriginalEndpoint?: string;
    afterUri?: string;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptSendToEndpointDefinition>) {
        super('InterceptSendToEndpointDefinition');
        Object.assign(this, init);
    }
}
// ...
export class OnCompletionDefinition extends WorkflowElement {
    stepName?: string = 'onCompletion';
    id?: string = 'onCompletion-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    mode?: string;
    onCompleteOnly?: boolean;
    onFailureOnly?: boolean;
    parallelProcessing?: boolean;
    executorService?: string;
    useOriginalMessage?: boolean;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<OnCompletionDefinition>) {
        super('OnCompletionDefinition');
        Object.assign(this, init);
    }
}
// ...
export class RouteConfigurationDefinition extends WorkflowElement {
    stepName?: string = 'routeConfiguration';
    id?: string;
    description?: string;
    errorHandler?: ErrorHandlerDefinition;
    intercept?: InterceptDefinition[] = [];
    interceptFrom?: InterceptFromDefinition[] = [];
    interceptSendToEndpoint?: InterceptSendToEndpointDefinition[] = [];
    onException?: OnExceptionDefinition[] = [];
    onCompletion?: OnCompletionDefinition[] = [];
    precondition?: string;
    public constructor(init?: Partial<RouteConfigurationDefinition>) {
        super('RouteConfigurationDefinition');
        Object.assign(this, init);
    }
}

export class RouteContextRefDefinition extends WorkflowElement {
    stepName?: string = 'routeContextRef';
    ref: string = '';
    public constructor(init?: Partial<RouteContextRefDefinition>) {
        super('RouteContextRefDefinition');
        Object.assign(this, init);
    }
}

export class RouteDefinition extends WorkflowElement {
    stepName?: string = 'route';
    id?: string = 'route-' + uuidv4().substring(0,4);
    description?: string;
    group?: string;
    nodePrefixId?: string;
    routeConfigurationId?: string;
    autoStartup?: boolean;
    startupOrder?: number;
    streamCache?: boolean;
    trace?: boolean;
    messageHistory?: boolean;
    logMask?: boolean;
    errorHandlerRef?: string;
    shutdownRoute?: string;
    shutdownRunningTask?: string;
    precondition?: string;
    errorHandler?: ErrorHandlerDefinition;
    inputType?: InputTypeDefinition;
    outputType?: OutputTypeDefinition;
    from: FromDefinition = new FromDefinition();
    public constructor(init?: Partial<RouteDefinition>) {
        super('RouteDefinition');
        Object.assign(this, init);
    }
}
// ...
export class DeleteDefinition extends WorkflowElement {
    stepName?: string = 'delete';
    id?: string = 'delete-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<DeleteDefinition>) {
        super('DeleteDefinition');
        Object.assign(this, init);
    }
}

export class GetDefinition extends WorkflowElement {
    stepName?: string = 'get';
    id?: string = 'get-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<GetDefinition>) {
        super('GetDefinition');
        Object.assign(this, init);
    }
}

export class HeadDefinition extends WorkflowElement {
    stepName?: string = 'head';
    id?: string = 'head-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<HeadDefinition>) {
        super('HeadDefinition');
        Object.assign(this, init);
    }
}
// ...
export class OpenApiDefinition extends WorkflowElement {
    stepName?: string = 'openApi';
    id?: string = 'openApi-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    specification: string = '';
    routeId?: string;
    missingOperation?: string;
    mockIncludePattern?: string;
    public constructor(init?: Partial<OpenApiDefinition>) {
        super('OpenApiDefinition');
        Object.assign(this, init);
    }
}
// ...
export class PatchDefinition extends WorkflowElement {
    stepName?: string = 'patch';
    id?: string = 'patch-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<PatchDefinition>) {
        super('PatchDefinition');
        Object.assign(this, init);
    }
}

export class PostDefinition extends WorkflowElement {
    stepName?: string = 'post';
    id?: string = 'post-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<PostDefinition>) {
        super('PostDefinition');
        Object.assign(this, init);
    }
}

export class PutDefinition extends WorkflowElement {
    stepName?: string = 'put';
    id?: string = 'put-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    to?: string;
    param?: ParamDefinition[] = [];
    responseMessage?: ResponseMessageDefinition[] = [];
    security?: SecurityDefinition[] = [];
    consumes?: string;
    produces?: string;
    type?: string;
    outType?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    deprecated?: boolean;
    streamCache?: boolean;
    routeId?: string;
    public constructor(init?: Partial<PutDefinition>) {
        super('PutDefinition');
        Object.assign(this, init);
    }
}
// ...
export class RestDefinition extends WorkflowElement {
    stepName?: string = 'rest';
    id?: string = 'rest-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    path?: string;
    consumes?: string;
    produces?: string;
    bindingMode?: string;
    skipBindingOnErrorCode?: boolean;
    clientRequestValidation?: boolean;
    enableCORS?: boolean;
    enableNoContentResponse?: boolean;
    apiDocs?: boolean;
    tag?: string;
    openApi?: OpenApiDefinition;
    securityDefinitions?: RestSecuritiesDefinition;
    securityRequirements?: SecurityDefinition[] = [];
    delete?: DeleteDefinition[] = [];
    get?: GetDefinition[] = [];
    head?: HeadDefinition[] = [];
    patch?: PatchDefinition[] = [];
    post?: PostDefinition[] = [];
    put?: PutDefinition[] = [];
    public constructor(init?: Partial<RestDefinition>) {
        super('RestDefinition');
        Object.assign(this, init);
    }
}

export class RestPropertyDefinition extends WorkflowElement {
    stepName?: string = 'restProperty';
    key: string = '';
    value: string = '';
    public constructor(init?: Partial<RestPropertyDefinition>) {
        super('RestPropertyDefinition');
        Object.assign(this, init);
    }
}

export class RestSecuritiesDefinition extends WorkflowElement {
    stepName?: string = 'securityDefinitions';
    apiKey?: ApiKeyDefinition;
    basicAuth?: BasicAuthDefinition;
    bearer?: BearerTokenDefinition;
    mutualTLS?: MutualTLSDefinition;
    oauth2?: OAuth2Definition;
    openIdConnect?: OpenIdConnectDefinition;
    public constructor(init?: Partial<RestSecuritiesDefinition>) {
        super('RestSecuritiesDefinition');
        Object.assign(this, init);
    }
}

export class RestsDefinition extends WorkflowElement {
    stepName?: string = 'rests';
    id?: string = 'rests-' + uuidv4().substring(0,4);
    description?: string;
    rest?: RestDefinition[] = [];
    public constructor(init?: Partial<RestsDefinition>) {
        super('RestsDefinition');
        Object.assign(this, init);
    }
}

export class SecurityDefinition extends WorkflowElement {
    stepName?: string = 'security';
    key: string = '';
    scopes?: string;
    public constructor(init?: Partial<SecurityDefinition>) {
        super('SecurityDefinition');
        Object.assign(this, init);
    }
}